fix: 完成P0构建问题修复 - 解决antd图标兼容性和页面导出问题
- 安装兼容版本antd@5.20.6和@ant-design/icons@5.4.0
- 配置Next.js transpilePackages解决模块兼容性问题
- 将页面组件改为动态导入避免SSR问题
- 创建页面组件文件分离客户端代码
- 验证构建成功和基础测试通过
