+++ package.json
@@ -16,13 +16,13 @@
     "db:seed": "node scripts/seed.js"
   },
   "dependencies": {
-    "@ant-design/icons": "^6.0.0",
+    "@ant-design/icons": "^5.4.0",
     "@heroicons/react": "^2.2.0",
     "@tailwindcss/forms": "^0.5.10",
     "@tailwindcss/typography": "^0.5.16",
     "@types/jsonwebtoken": "^9.0.10",
     "@types/react-grid-layout": "^1.3.5",
-    "antd": "^5.26.7",
+    "antd": "^5.20.6",
     "bcryptjs": "^2.4.3",
     "better-sqlite3": "^11.10.0",
     "dayjs": "^1.11.13",
